Refactor WizardScreen to use ViewModel
Introduces `WizardViewModel` to manage the state and logic of the `WizardScreen`.
This change centralizes state management and improves the testability of the screen.

Refactor WizardScreen to use ViewModel

Refactor WizardScreen to use ViewModel

Introduces `WizardViewModel` to manage the state and logic of the `WizardScreen`.
This change centralizes state management and improves the testability of the screen.